Rename src/ directory
