+++ Makefile
@@ -16,8 +16,8 @@ install: test
 	mkdir -p ${targetdir}/${libexecdir}/igcollect
 	mkdir -p ${targetdir}/${libexecdir}/igcollect/libigcollect
 	mkdir -p ${targetdir}/share/java
-	install src/*.py		${targetdir}/${libexecdir}/igcollect
-	install src/libigcollect/*.py	${targetdir}/${libexecdir}/igcollect/libigcollect
+	install igcollect/*.py		${targetdir}/${libexecdir}/igcollect
+	install igcollect/libigcollect/*.py	${targetdir}/${libexecdir}/igcollect/libigcollect
 	install share/java/*.jar	${targetdir}/share/java
 
 test:
