+++ igcollect/artfiles.py
