+++ igcollect/mysql.py
