+++ igcollect/pgsql_globals.py
