+++ igcollect/rabbitmq.py
