+++ igcollect/redis.py
