+++ igcollect/smart.py
