+++ igcollect/switch.py
